Address PR review and fix CI: move TimelineEntry, ARIA, prettier
Five follow-ups for the ChatMessage decomposition PR:

- Break the ChatMessage.tsx ↔ ChatMessage.styles.ts circular type
  import. Pull `TimelineEntry` into a new `chat/types.ts` module and
  update `ChatMessage.tsx`, `ChatMessage.styles.ts`, and
  `ChatMessageTimeline.tsx` to import from there. `ChatMessage.tsx`
  still re-exports `TimelineEntry` so the public API remains intact.
- Fix the dead conditional in `ApprovalIndicator` (both
  ternary branches resolved to `"3.5rem"` — collapse to a literal).
- `ChatMessageSourcePreview`: replace the array-index React key on the
  `SourceItem` map with `${messageId}-source-${index}` so reorder/insert
  diffs don't re-mount unaffected items.
- `ChatMessageSourcePreview`: add ARIA semantics to the LabelMenu
  dropdown (`role="menu"`, `aria-label`, `aria-haspopup="menu"` and
  `aria-expanded` on the trigger), matching the keyboard-accessible
  pattern already used by `ToolUsageIndicator`.
- Run `yarn run prettier --write` over the chat dir to fix the lint
  failure (`ChatMessageSourcePreview.tsx` was the unformatted file
  flagged by CI).
